+++ index.js
@@ -19,6 +19,7 @@ exports.getDirTree = getDirTree;
 
 function boot(options) {
 
+    options.listen = options.listen !== false;
     options.express = options.express || express;
     options.app = options.app || options.express();
     options.PORT = options.PORT || 3000;
@@ -64,6 +65,10 @@ function loadAfterBoot(options, dirTree) {
     exports.models = modelLoader(modelPaths, exports.services, options);
     exports.controllers = loadControllers(exports.models, dirTree, exports.services, options);
 
+    if (!options.listen) {
+        return;
+    }
+
     options.verbose && exports.services && (console.log('\n\nServices available:'), Object.keys(exports.services).forEach(service => console.log(`    ${service}`)));
     options.verbose && exports.models && (console.log('\n\nModels available:'), Object.keys(exports.models).forEach(model => console.log(`    ${model}`)));
     options.verbose && exports.controllers && (console.log('\n\nControllers available:'), Object.keys(exports.controllers).forEach(controller => console.log(`    ${controller}`)));
@@ -71,6 +76,8 @@ function loadAfterBoot(options, dirTree) {
     // Start node app
     options.app.set('view engine', options.viewEngine);
     options.app.set('views', dirTree.views);
+
+    // TODO: this should be in the docs and optional
     // parse application/x-www-form-urlencoded
     options.app.use(bodyParser.urlencoded({ extended: false }));
     // parse application/json